Move cdp interfaces to connection
